feat : 카테고리 store 값 초기화
- 로그인후 처음 화면 랜더링시 카테고리 값을 서버로부터 받아와 초기화함
